refactor: Dependency Inversion
